+++ nemo/collections/llm/api.py
@@ -58,7 +58,7 @@ from nemo.lightning import (
 from nemo.lightning.base import NEMO_MODELS_CACHE
 from nemo.lightning.ckpt_utils import ckpt_to_context_subdir
 from nemo.lightning.pytorch.callbacks import PEFT, JitTransform, ModelTransform
-from nemo.lightning.callback_group import CallbackGroup
+from nemo.utils.callbacks.callback_group import CallbackGroup
 from nemo.utils import logging
 from nemo.utils.get_rank import is_global_rank_zero
