+++ nemo/utils/callbacks/__init__.py
@@ -15,3 +15,6 @@
 from nemo.utils.callbacks.cuda_graph import CUDAGraphCallback
 from nemo.utils.callbacks.nemo_model_checkpoint import NeMoModelCheckpoint
 from nemo.utils.callbacks.preemption import PreemptionCallback
+from nemo.utils.callbacks.base_callback import BaseCallback
+from nemo.utils.callbacks.one_logger_callback import OneLoggerNeMoCallback
+from nemo.utils.callbacks.callback_group import CallbackGroup, hook_class_init_with_callbacks
