+++ nemo/utils/callbacks/one_logger_callback.py
@@ -23,7 +23,7 @@ from nv_one_logger.training_telemetry.api.config import TrainingTelemetryConfig
 from nv_one_logger.training_telemetry.api.training_telemetry_provider import TrainingTelemetryProvider
 from nv_one_logger.training_telemetry.integration.pytorch_lightning import TimeEventCallback as OneLoggerPTLCallback
 
-from nemo.lightning.base_callback import BaseCallback
+from nemo.utils.callbacks.base_callback import BaseCallback
 from nemo.utils.meta_info_manager import (
     enable_one_logger,
     get_nemo_v1_callback_config,
@@ -63,3 +63,5 @@ class OneLoggerNeMoCallback(OneLoggerPTLCallback, BaseCallback):
             config = get_nemo_v1_callback_config(trainer=trainer)
         training_telemetry_config = TrainingTelemetryConfig(**config)
         TrainingTelemetryProvider.instance().set_training_telemetry_config(training_telemetry_config)
+
+
